feat: add outlet
