+++ news-api/src/main.tsx
@@ -1,13 +1,14 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import './index.css';
-import { Router, store } from './app';
+import { store } from './app';
 import { Provider } from 'react-redux';
+import App from './app/App.tsx';
 
 createRoot(document.getElementById('root')!).render(
     <StrictMode>
         <Provider store={store}>
-            <Router />
+            <App />
         </Provider>
     </StrictMode>
 );
