+++ news-api/src/shared/layouts/MainLayout.tsx
@@ -1,13 +1,10 @@
-import type { FC, ReactNode } from 'react';
+import type { FC } from 'react';
 import { useTheme } from '../lib';
+import { Footer, Header } from '../../widgets';
+import { Outlet } from 'react-router';
+import { UserTabs } from '../../widgets/UserTabs/UserTabs.tsx';
 
-type Props = {
-    header?: ReactNode;
-    children: ReactNode;
-    footer?: ReactNode;
-};
-
-export const MainLayout: FC<Props> = ({ children, header, footer }) => {
+export const MainLayout: FC = () => {
     const { theme } = useTheme();
 
     return (
@@ -20,9 +17,10 @@ export const MainLayout: FC<Props> = ({ children, header, footer }) => {
                 flexDirection: 'column',
             }}
         >
-            {header}
-            {children}
-            {footer}
+            <Header />
+            <UserTabs />
+            <Outlet />
+            <Footer />
         </div>
     );
 };
